requirements.txt
# Web scraping dependencies
requests==2.31.0
beautifulsoup4==4.12.2

# LLM API integration
openai==1.51.0

# Database - sqlite3 is built into Python standard library

# Web application
flask==3.0.0

# Environment variables
python-dotenv==1.0.0

# Testing
pytest==7.4.3

# Logging (built into Python standard library)